Moved much of logic out of command and into its own file.
